test(ai): isolate retry sleep lifecycle coverage
