+++ packages/ai/src/providers/openai-chatgpt-responses.test.ts
@@ -4,7 +4,6 @@ import { zstdDecompressSync } from "node:zlib";
 import { MAX_TIMER_TIMEOUT_MS } from "@openclaw/normalization-core/number-coercion";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { configureAiTransportHost } from "../host.js";
-import { sleepWithAbort } from "../internal/retry-sleep.js";
 import type { Context, Model } from "../types.js";
 import { SYSTEM_PROMPT_CACHE_BOUNDARY } from "../utils/system-prompt-cache-boundary.js";
 import {
@@ -79,62 +78,6 @@ function completedSseResponse(responseId = "resp_test"): Response {
     headers: { "content-type": "text/event-stream" },
   });
 }
-
-describe("sleepWithAbort listener lifecycle", () => {
-  afterEach(() => {
-    vi.restoreAllMocks();
-  });
-
-  it("unregisters the abort listener after a successful wait", async () => {
-    const controller = new AbortController();
-    const addSpy = vi.spyOn(controller.signal, "addEventListener");
-    const removeSpy = vi.spyOn(controller.signal, "removeEventListener");
-
-    await sleepWithAbort(5, controller.signal);
-
-    expect(addSpy).toHaveBeenCalledWith("abort", expect.any(Function), { once: true });
-    const onAbort = addSpy.mock.calls.find((call) => call[0] === "abort")?.[1];
-    expect(onAbort).toEqual(expect.any(Function));
-    expect(removeSpy).toHaveBeenCalledWith("abort", onAbort);
-  });
-
-  it("does not accumulate listeners across multiple successful waits on one signal", async () => {
-    const controller = new AbortController();
-    const addSpy = vi.spyOn(controller.signal, "addEventListener");
-    const removeSpy = vi.spyOn(controller.signal, "removeEventListener");
-
-    await sleepWithAbort(1, controller.signal);
-    await sleepWithAbort(1, controller.signal);
-    await sleepWithAbort(1, controller.signal);
-
-    const abortAdds = addSpy.mock.calls.filter((call) => call[0] === "abort");
-    const abortRemoves = removeSpy.mock.calls.filter((call) => call[0] === "abort");
-    expect(abortAdds).toHaveLength(3);
-    expect(abortRemoves).toHaveLength(3);
-    for (let i = 0; i < abortAdds.length; i += 1) {
-      expect(abortRemoves[i]?.[1]).toBe(abortAdds[i]?.[1]);
-    }
-  });
-
-  it("rejects when the signal aborts during the wait and cleans up the timeout", async () => {
-    const controller = new AbortController();
-    const clearTimeoutSpy = vi.spyOn(globalThis, "clearTimeout");
-    const removeSpy = vi.spyOn(controller.signal, "removeEventListener");
-    const pending = sleepWithAbort(60_000, controller.signal);
-    queueMicrotask(() => {
-      controller.abort();
-    });
-    await expect(pending).rejects.toThrow("Request was aborted");
-    expect(clearTimeoutSpy).toHaveBeenCalledOnce();
-    expect(removeSpy).toHaveBeenCalledWith("abort", expect.any(Function));
-  });
-
-  it("rejects immediately when the signal is already aborted", async () => {
-    const controller = new AbortController();
-    controller.abort();
-    await expect(sleepWithAbort(5, controller.signal)).rejects.toThrow("Request was aborted");
-  });
-});
 
 describe("extractOpenAICodexAccountId", () => {
   it("decodes URL-safe base64 JWT payloads", () => {
